refactor(fs-watch): split rename handling out of fsChangeHandlers
The previous commit's seven-line header update took the file from 294 to 301,
one over the limit. It had been sitting one line under, so the next change of
any kind was going to break it — trimming the comment would have left the same
trap for whoever came next.

The seam is the one the code already had: a rename is the only event kind that
MUTATES the path map every other event reads, which is why it is dispatched
first, one at a time, against a freshly-built map each time. applyRenamePair,
handleRenameEvent, readAndRouteOrMarkMissing and dispatchRenames move to
fsRenameHandlers.ts; fsChangeHandlers keeps remove, modify/create, media and
the batch dispatcher. 139 + 181 lines. All 180 tests pass unchanged.

Worth recording how this got through: predelta was green, then I committed a
docs-only change on top and did not re-run it. A comment is not a safe edit in
a file measured by line count. Also, `pnpm check:all > log; echo exit=$? >> log`
backgrounded reports the ECHO's exit status to the harness, so the run looked
green from outside while the log said 1 — the number that matters is the one
inside the file, which is why it is written there.
